requirements.txt
django
djangorestframework
faiss-cpu
langchain
langchain-community
sentence-transformers
numpy